moved app files into Src folder
