+++ DAQ_System/Core/Src/main.c
@@ -27,7 +27,7 @@
 
 /* Private includes ----------------------------------------------------------*/
 /* USER CODE BEGIN Includes */
-#include "../../Program/app.hpp"
+#include "../../Program/Src/app.hpp"
 #include "retarget.h"
 /* USER CODE END Includes */
